requirements.txt
# FastAPI e dependências
fastapi==0.115.7
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# AWS
boto3==1.29.0
botocore==1.32.0

# Processamento de dados
python-dateutil==2.8.2
pydantic==2.7.4
typing-extensions==4.8.0

# Utilidades
python-dotenv==1.0.0
pytz==2023.3.post1

# Segurança
cryptography==41.0.5
python-jose[cryptography]==3.3.0

# Gerenciamento de Arquivos
aiofiles==23.2.1